+++ CHANGELOG.md
@@ -156,5 +156,6 @@
 
 # 0.2.9
 * Replace `Int64` in `Send`, `Forward`, `Edit`, `Delete`, `Stop` message's `Persistable` methods on `ID Chat`
+* Replace `Int` in `Forward`, `Edit`, `Delete`, `Stop` message's `Persistable` methods on `ID Message`
 * Replace identifiers in `Callback`, `Message`, `Sender`, `Update` on specified `ID` data family instances
 * Move chat identifiers from `Channel` and `Group` to `Origin` and `Moving`, delete `Conversation` module